# Optimize conv2d pw quantized